+++ html/evaluacion.html
@@ -79,6 +79,7 @@
     integrity="sha512-894YE6QWD5I59HgZOGReFYm4dnWc1Qt5NtvYSaNcOP+u1T9qYdvdihz0PPSiiqn/+/3e7Jo4EaG7TubfWGUrMQ=="
     crossorigin="anonymous" referrerpolicy="no-referrer"></script>
   <script src="../js/evaluacion.js" type="module"></script>
+  <script src="../js/respuestas.js"></script>
 </body>
 
 </html>
