--- js/ejemplo.js
@@ -1,3 +1,0 @@
-function evaluacion(){
-    console.log("hola mundo");
-}
